e-commerce/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multipledispatch==1.0.0
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
zope.event==5.0
zope.interface==7.0.3
celery
django-redis
